requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.1
colorama==0.4.5
decorator==4.4.2
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
moviepy==1.0.3
numpy==1.23.2
Pillow==9.2.0
praw==7.6.0
prawcore==2.3.0
proglog==0.1.10
python-dotenv==0.20.0
requests==2.28.1
tqdm==4.64.0
update-checker==0.18.0
urllib3==1.26.12
websocket-client==1.3.3
